Back-end/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.8.5
langchain>=0.0.312
websockets==12.0
pydantic>=2.4.2
python-multipart==0.0.6
httpx==0.25.0 